requirements.txt
amqp==2.5.2
anyjson==0.3.3
astroid==2.1.0
beautifulsoup4==4.6.0
billiard==3.6.1.0
cachetools==3.1.0
celery==4.3.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.4
defusedxml==0.5.0
Django==2.1.15
django-allauth==0.32.0
django-anymail==6.0
django-appconf==1.0.2
django-ckeditor==5.6.1
django-cleanup==3.0.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-imagekit==4.0.2
django-js-asset==1.1.0
django-material==1.5.3
django-model-utils==4.0.0
django-semanticui-forms==1.6.5
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.3
easy-thumbnails==2.6
google-auth==1.6.3
html5lib==1.0b10
httplib2==0.12.3
idna==2.8
importlib-metadata==1.3.0
isort==4.2.15
kombu==4.6.7
lazy-object-proxy==1.3.1
mailjet-rest==1.3.3
Markdown==3.0.1
mccabe==0.6.1
more-itertools==8.0.2
oauthlib==3.0.1
pilkit==2.0
Pillow==6.2.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyflakes==1.5.0
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.9
redis==3.3.11
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.3.2
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
webencodings==0.5.1
wrapt==1.11.1
xlrd==1.2.0
zipp==0.6.0
